This is the code that works for Linux
I am trying to convert it to MAC, but keep getting a few errors when I run it. Like on line 5 it cannot call name "image_read" and i have been unable to figure out a solution to that problem. All help is appreciated as I am new to all this!